refactor reusable code to ensure aws-sdk-go repository
