+++ cmd/ack-generate/command/release.go
@@ -24,6 +24,7 @@ import (
 
 	ackgenerate "github.com/aws-controllers-k8s/code-generator/pkg/generate/ack"
 	ackmetadata "github.com/aws-controllers-k8s/code-generator/pkg/metadata"
+	"github.com/aws-controllers-k8s/code-generator/pkg/sdk"
 )
 
 var optReleaseOutputPath string
@@ -58,11 +59,13 @@ func generateRelease(cmd *cobra.Command, args []string) error {
 	// version supplied hasn't been used (as a Git tag) before...
 	releaseVersion := strings.ToLower(args[1])
 
-	ctx, cancel := contextWithSigterm(context.Background())
+	ctx, cancel := sdk.ContextWithSigterm(context.Background())
 	defer cancel()
-	if err := ensureSDKRepo(ctx, optCacheDir, optRefreshCache); err != nil {
+	sdkDirPath, err := sdk.EnsureRepo(ctx, optCacheDir, optRefreshCache, optAWSSDKGoVersion, optOutputPath)
+	if err != nil {
 		return err
 	}
+	sdkDir = sdkDirPath
 	m, err := loadModel(svcAlias, "", "", ackgenerate.DefaultConfig)
 	if err != nil {
 		return err
@@ -93,7 +96,7 @@ func generateRelease(cmd *cobra.Command, args []string) error {
 		}
 		outPath := filepath.Join(optReleaseOutputPath, path)
 		outDir := filepath.Dir(outPath)
-		if _, err := ensureDir(outDir); err != nil {
+		if _, err := sdk.EnsureDir(outDir); err != nil {
 			return err
 		}
 		if err = ioutil.WriteFile(outPath, contents.Bytes(), 0666); err != nil {
